requirements.txt
# 🚀 Universal SDK-to-MCP Converter Dependencies

# Core MCP Framework
mcp>=1.9.0

# LLM Auto-Configuration (Phase 6)
openai>=1.100.0
python-dotenv>=1.1.0

# Plugin System
PyYAML>=6.0

# Core Python Support
typing-extensions>=4.0.0

# Example/Test SDKs (install as needed)
requests>=2.31.0

# Optional: Popular SDKs for testing
# PyGithub>=2.0.0              # GitHub SDK
# kubernetes>=28.0.0           # Kubernetes SDK  
# boto3>=1.34.0               # AWS SDK
# stripe>=7.0.0               # Stripe SDK

# Development Tools (optional)
# pytest>=7.0.0
# black>=23.0.0